Fully move HTML module
